Move top-level functions to Api.php
